hash: move files & use name convention
