+++ src/typeorm/entities/users/Graphql.ts
@@ -1,8 +1,8 @@
 import { gql } from 'apollo-server';
-import HashProvider from './providers/HashProvider/implementations/HashProvider';
 import { getRepository } from 'typeorm';
 import User from './User-entity';
 import JWTProvider from './providers/jwt-provider/jwt-provider';
+import HashProvider from './providers/hash-provider/hash-provider';
 
 export const userTypeDefs = gql`
   type User {
